Re-factor: form and preview split apart
