+++ app/frontend/test/unit/specs/submissions/SubmissionsHome.spec.js
@@ -1,8 +1,9 @@
-import { shallowMount, createLocalVue } from '@vue/test-utils'
+import { shallowMount, createLocalVue, mount } from '@vue/test-utils'
 import Vuex from 'vuex'
 import VueRouter from 'vue-router'
 import SubmissionsHome from '@/submissions/views/SubmissionsHome.vue'
 import Yield from '@/submissions/components/SubmissionForm/Yield.vue'
+// import ActivitySubmissionForm from '@/submissions/components/SubmissionForm/ActivitySubmissionForm.vue'
 
 import { FETCH_CODES } from '@/submissions/store/actions.types.js'
 
@@ -56,12 +57,12 @@ describe('SubmissionsHome.vue', () => {
   })
 
   it('renders components based on the activity selected (Yield exists on construction report', (done) => {
-    const wrapper = shallowMount(SubmissionsHome, {
+    const wrapper = mount(SubmissionsHome, {
       localVue,
       store,
       router
     })
-    wrapper.setData({ activityType: 'CON', formIsFlat: true })
+    wrapper.setData({ activityType: 'CON' })
     wrapper.vm.$nextTick(() => {
       expect(wrapper.find(Yield).exists()).toBe(true)
       done()
